Fix cylinder inertia tensor computation.
