+++ src/dynamics/mod.rs
@@ -23,6 +23,7 @@ mod mass_properties;
 mod mass_properties_ball;
 mod mass_properties_capsule;
 mod mass_properties_cuboid;
+mod mass_properties_cylinder;
 #[cfg(feature = "dim2")]
 mod mass_properties_polygon;
 mod rigid_body;
